streamlit-app/requirements.txt
pandas
streamlit-option-menu
seaborn
matplotlib
numpy
scikit-learn
pathlib
statsmodels